Refactor: ui package 생성 및 클래스 위치 이동
